backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
pypdf
langchain
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-cloud-storage
google-cloud-vision
fpdf
firebase-admin